+++ packages/dart_lints/example/example_app/bin/example_app.dart
@@ -1,0 +1,5 @@
+import 'package:example_app/always_declare_return_types.dart' as example_app;
+
+void main(List<String> arguments) {
+  print('Hello world: ${example_app.calculate()}!');
+}