+++ packages/dart_lints/example/example_app/lib/always_declare_return_types.dart
@@ -1,0 +1,4 @@
+// LINT: always_declare_return_types
+calculate() {
+  return 6 * 7;
+}