+++ packages/dart_lints/example/example_app/test/example_app_test.dart
@@ -1,0 +1,8 @@
+import 'package:example_app/always_declare_return_types.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('calculate', () {
+    expect(calculate(), 42);
+  });
+}