+++ packages/dart_lints/lib/constants.dart
@@ -1,0 +1,5 @@
+const kPackageId = 'dart_lints';
+
+const _kDartBaseUrl = 'https://dart-lang.github.io/linter/lints/';
+
+final kUri = Uri.parse(_kDartBaseUrl);