+++ packages/dart_lints/lib/src/always_declare_return_types.dart
@@ -1,0 +1,55 @@
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/token.dart';
+import 'package:sidecar/sidecar.dart';
+
+import '../constants.dart';
+
+class AlwaysDeclareReturnTypes extends Rule with Lint {
+  static const _id = 'always_declare_return_types';
+
+  @override
+  LintCode get code =>
+      LintCode(_id, package: kPackageId, url: kUri.resolve(_id));
+
+  @override
+  void initializeVisitor(NodeRegistry registry) {
+    registry
+      ..addFunctionDeclaration(this)
+      ..addFunctionTypeAlias(this)
+      ..addMethodDeclaration(this);
+  }
+
+  @override
+  void visitFunctionDeclaration(FunctionDeclaration node) {
+    if (!node.isSetter && node.returnType == null) {
+      reportToken(
+        node.name2,
+        message:
+            "The function ${node.name2.lexeme} should have a return type but doesn't.",
+        correction: 'Try adding a return type to the function.',
+      );
+    }
+  }
+
+  @override
+  void visitFunctionTypeAlias(FunctionTypeAlias node) {
+    if (node.returnType == null) {
+      reportToken(node.name2,
+          message:
+              "The function ${node.name2.lexeme} should have a return type but doesn't.",
+          correction: 'Try adding a return type to the function.');
+    }
+  }
+
+  @override
+  void visitMethodDeclaration(MethodDeclaration node) {
+    if (!node.isSetter &&
+        node.returnType == null &&
+        node.name2.type != TokenType.INDEX_EQ) {
+      reportToken(node.name2,
+          message:
+              "The method ${node.name2.lexeme} should have a return type but doesn't.",
+          correction: 'Try adding a return type to the method.');
+    }
+  }
+}