+++ packages/dcm_lints/example/dcm_lints_example.dart
@@ -1,0 +1,6 @@
+// import 'package:dcm_lints/dcm_lints.dart';
+
+// void main() {
+//   var awesome = Awesome();
+//   print('awesome: ${awesome.isAwesome}');
+// }