+++ packages/design_system_lints/lib/src/avoid_box_shadow_literal.dart
@@ -4,31 +4,25 @@ import 'package:sidecar/sidecar.dart';
 
 import 'constants.dart';
 
-class AvoidBoxShadowLiteral extends LintRule with LintVisitor {
-  @override
-  String get code => 'avoid_box_shadow_literal';
+/// Avoid hardcoding BoxShadows.
+class AvoidBoxShadowLiteral extends Rule with Lint {
+  static const _id = 'avoid_box_shadow_literal';
+  static const _message = 'Avoid BoxShadow literal';
+  static const _correction = 'Use values in design system spec instead';
 
   @override
-  String get packageName => kDesignSystemPackageId;
+  LintCode get code => LintCode(_id, package: kPackageId, url: kUrl);
 
   @override
-  String get url => kUrl;
+  void initializeVisitor(NodeRegistry registry) {
+    registry.addInstanceCreationExpression(this);
+  }
 
-  @override
-  SidecarAstVisitor Function() get visitorCreator => _Visitor.new;
-}
-
-class _Visitor extends SidecarAstVisitor {
   @override
   void visitInstanceCreationExpression(InstanceCreationExpression node) {
     final element = node.constructorName.staticElement;
     if (boxShadowType.isAssignableFromType(element?.returnType)) {
-      reportAstNode(
-        node,
-        message: 'Avoid BoxShadow literal',
-        correction: 'Use design system spec instead.',
-      );
+      reportAstNode(node, message: _message, correction: _correction);
     }
-    super.visitInstanceCreationExpression(node);
   }
 }