+++ packages/design_system_lints/test/design_system_lints_test.dart
@@ -1,6 +1,0 @@
-import 'package:design_system_lints/design_system_lints.dart';
-import 'package:test/test.dart';
-
-void main() {
-  //
-}