+++ packages/intl_lints/lib/src/constants.dart
@@ -1,0 +1,4 @@
+const packageId = 'intl_lints';
+final _kUrl =
+    'https://github.com/pattobrien/lints/tree/main/packages/intl_lints/';
+final kUri = Uri.parse(_kUrl);