+++ packages/design_system_lints/CHANGELOG.md
@@ -1,3 +1,21 @@
-## 1.0.0
+## 0.1.0-dev.6
+
+- chore: Updated to Sidecar 0.1.0-dev.11 APIs.
+
+## 0.1.0-dev.5
+
+- docs: added API documentation to rule classes
+- chore: Updated to Sidecar 0.1.0-dev.9 APIs.
+
+## 0.1.0-dev.4
+
+- docs: added API documentation to rule classes
+- chore: cleaned up static analysis lints for unused members, imports, etc.
+
+## 0.1.0-dev.3
+
+- chore: Utilize new Sidecar v0.1.0.-dev.7 APIs
+
+## 0.1.0-dev.1
 
 - Initial version.